this is the fastest map i can get for now: learnt that integer equality is faster than memcmp, even inlined, and even by paying the cost of a runtime  on key_typesize; remains to be seen if it can be improved with avx

+++ src/ion/modules/map/functions/map_hash.h
@@ -1,4 +1,4 @@
-static inline u32 map_hash (
+static inline u64 map_hash (
     void* key,
     u32 key_typesize
 );
