+++ src/ion/modules/map/functions/map_init.h
@@ -5,6 +5,6 @@ static inline
 struct map map_init (
     u32 key_typesize,
     u32 value_typesize,
-    u32 initial_capacity,
+    u64 initial_capacity,
     struct memory* allocator
 );
